pyproject.toml
[project]
name = "ponderadaa"
version = "0.1.0"
description = ""
authors = [
    {name = "isarocha04",email = "isabela.rocha@sou.inteli.edu.br"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "supabase (>=2.13.0,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)"
]

[tool.poetry]
packages = [{include = "ponderadaa", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


